Add function to set up loggers
Create an init_loggers() function in main to initialize two loggers
for use in the program. The first logger prints to the console and
is meant for logging of program runtime information, useful to us.
The other log is meant to log all git interactions done by the
user to a file, for the use of the user.

Implement file path operations in an os-agnostic way and gracefully
recover from errors related to file path operation.

Fix various minor code errors

Clean up formatting and fix spelling errors made in code